% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_simulations.R
\name{assign_column_names}
\alias{assign_column_names}
\title{Helper function to assign column names to a matrix}
\usage{
assign_column_names(x)
}
\arguments{
\item{x}{the matrix to which column names will be assigned}
}
\description{
Helper function to assign column names to a matrix
}
