% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_selection.R
\name{bestsubsetfit}
\alias{bestsubsetfit}
\title{Fit a linear regression model with best subset selection and backward
elimination method.}
\usage{
bestsubsetfit(
  x,
  y,
  method = c("exhaustive", "backward"),
  foldid = NULL,
  nfolds = 10,
  criterion = c("cv", "aic", "bic"),
  shrinkage = c("none", "pwsf", "global", "breiman"),
  choice = c("tenfold", "loocv"),
  standardize = FALSE,
  type.measure = c("mse", "mae"),
  lower.limits = -Inf,
  upper.limits = Inf,
  nonnegative = FALSE,
  lambda.min.ratio = 1e-08,
  sigma2 = NULL,
  betatypes = NULL
)
}
\arguments{
\item{x}{The standardized matrix of predictors.}

\item{y}{A numeric centered quantitative vector of response.}

\item{method}{Specifies which method to use for variable selection: either
exhaustive search or backward elimination method. The default is exhaustive
search via leaps and bound algorithm. See `leaps` package in R for details.}

\item{foldid}{A vector of values between 1 and nfolds identifying which fold
each observation is in while conducting cross-validation. Used when criterion
is cross-validation (cv) or choice is tenfold. It can be missing when nfolds
is provided.}

\item{nfolds}{Number of folds for cross-validation. Default is 10.}

\item{criterion}{Specifies which criterion to use when selecting the final
model. Default is cross-validation. Other available options are AIC and BIC.}

\item{shrinkage}{Specifies which shrinkage method of parameter estimates is
required. Three options are available. The default is `none`, and no shrinkage
is conducted. Parameter-wise shrinkage factors, global, or Breiman
method can be used to estimate shrinkage factors.}

\item{choice}{Specifies which method of estimating shrinkage factors to use:
tenfold cross-validation or leave-one-out cross-validation. Default is tenfold cv.}

\item{standardize}{Specifies whether to standardize the predictors matrix
\code{`x`} to have mean 0 and unit variance. If set to TRUE, \code{`x`} will be
 standardized. Additionally, the response variable \code{`y`} will be centered
 by subtracting its mean. This standardization step can be useful to ensure
 that the predictors are on a comparable scale. By default,
 \code{standardize=FALSE}, indicating that no standardization will be performed.
 This assumes that users have already standardized their data.}

\item{type.measure}{Loss to use for cross-validation. Currently two options,
are available. The default is type.measure="mse", which uses squared-error.
Other option include type.measure = "mae" (mean absolute error). This
option is only applicable to the Breiman method.}

\item{lower.limits}{Vector of lower limits for each shrinkage factor.
The default is -Inf, and negative shrinkage factors will be estimated. If
\code{nonnegative = TRUE}, the program will set \code{lower.limits = 0}, and
nonnegative shrinkage factors will be estimated. In short, setting
\code{lower.limits = 0} is identical to setting \code{nonnegative = TRUE}.
This option is only applicable to the Breiman method.}

\item{upper.limits}{Vector of upper limits for the shrinkage factors.
The default is Inf, and unbounded shrinkage factors will be estimated. This
option is only applicable to the Breiman method}

\item{nonnegative}{Specify whether nonnegative shrinkage factors should be
estimated. The default is FALSE, implying thatnegative shrinkage factors may be
 estimated. See the `lower.limits` argument for instructions on estimating
 nonnegative shrinkage factors for the Breiman method.}

\item{lambda.min.ratio}{lambda can be provided if the user wants to specify
the lambda sequence, but typical usage is for the program to construct the
lambda sequence on its own. When automatically generated, the lambda sequence
is determined by lambda.max and lambda.min.ratio. The latter is the ratio of
smallest value of the generated lambda sequence (say lambda.min) to lambda.max.
The program generates nlambda values linear on the log scale from lambda.max
down to lambda.min. lambda.max is not user-specified but is computed from the
input x and y: it is the smallest value for lambda such that all the
coefficients are zero. Default is lambda.min.ratio = 1e-08.This option is only
applicable to the Breiman method.}

\item{sigma2}{Residual variance from the full model for calculation of AIC or
BIC. The user must supply if criterion is AIC or BIC.}

\item{betatypes}{Not used but added for consistency with the oracle_model().}
}
\value{
Returns a list with the following components:
\item{beta:}{Regression estimates of standardized variables. If shrinkage is
not "none," shrunken estimates are returned.}
\item{nvar:}{The number of variables selected.}
\item{shrinkagefactors:}{Shrinkage factors for the selected variables.}
\item{x:}{Standardized design matrix of the selected variables.}
\item{y:}{Centered response variable.}
\item{lambda,gamma:}{Parameters used in penalized methods, such as adaptive
lasso. It is not applicable in the current context. However, the function
returns the two parameters for consistency with the interface of penalized methods.
}
}
\description{
Conduct variable selection using best subset selection
(exhaustive search) and backward elimination. The final model can be chosen
using cross-validation, AIC, or BIC. The regression estimates
can be subjected to shrinkage, which is estimated using PWSF, global,
or the Breiman method. The function returns unshrunken or shrunken estimates.
}
