% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data_master.R
\name{beta_type}
\alias{beta_type}
\title{Generates true regression coefficients}
\usage{
beta_type(nvar = 15, type = c("a", "b", "c", "d"))
}
\arguments{
\item{nvar}{The number of variables. Default is 15}

\item{type}{The type of regression coefficients. See Table 1 of the reference}
}
\description{
Generates four types of regression coefficients
}
\references{
Kipruto, E. and Sauerbrei, W. (2022). Comparison of variable 
selection procedures and investigation of the role of shrinkage in linear
regression-protocol of a simulation study in low-dimensional data. Plos one, 
17(10), p.e0271240.
}
