% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgr.r
\name{bgr}
\alias{bgr}
\title{Bayesian group ridge regression}
\usage{
bgr(y, X, g, prior = "g", a = 0.5, b = 0.5, c = 1, mcmc = 5000L,
  burnin = 1000L, thin = 10L, verbose = TRUE, light = FALSE)
}
\arguments{
\item{y}{response vector of length n.}

\item{X}{n by p data matrix.}

\item{g}{vector of length p for group memberships.}

\item{prior}{character. Either "bp" or "g". See Details.}

\item{a}{hyperparameter.}

\item{b}{hyperparameter.}

\item{c}{hyperparameter.}

\item{mcmc}{integer. Number of desired samples.}

\item{burnin}{integer. Number of burn-in samples.}

\item{thin}{integer. Number of consecutive samples to skip.}

\item{verbose}{logical. Whether information on progress should be be printed.}

\item{light}{logical. If TRUE, only return a summary of the samples.}
}
\value{
An object of class \code{\link{list}}
}
\description{
This function implements a Gibbs algorithm for the ridge regression model with
grouped variables (predictors).
}
\details{
blablabla
}
\examples{
\dontrun{
# Generate data
set.seed(2019)
n <- 20
p <- 20
X <- matrix(rnorm(n*p), n, p)
trueBeta <- rep(0, p)
trueBeta[1:10] <- 1
y <- X\%*\%trueBeta + rnorm(n, sd=0.5)

# Run gibbs sampler
res <- br(y, X, prior="bp", a=0.5, b=0.5, mcmc=1000, burnin=1000, thin=10, verbose=TRUE)

# Extract summary
res$summary
}

}
\references{
Leday, G.G.R. et al. (2019)
}
\author{
Gwenael G.R. Leday
}
