% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brg.r
\name{brg}
\alias{brg}
\title{Bayesian linear regression with global shrinkage priors}
\usage{
brg(
  y,
  X,
  prior = "BetaPrime",
  a = 0.5,
  b = 0.5,
  mcmc = 5000L,
  burnin = 5000L,
  thin = 10L,
  verbose = TRUE,
  output = "both",
  BP = "GG",
  tau2_0 = 1
)
}
\arguments{
\item{y}{numeric response vector of length n.}

\item{X}{n by p data matrix.}

\item{prior}{character. See Details.}

\item{a}{hyperparameter.}

\item{b}{hyperparameter.}

\item{mcmc}{integer. Number of desired samples.}

\item{burnin}{integer. Number of burn-in samples.}

\item{thin}{integer. Number of consecutive samples to skip.}

\item{verbose}{logical. Whether information on progress should be printed.}

\item{output}{character. Either "samples", "summary" or "both".}

\item{BP}{character. Parametrization of Beta Prime prior. Either "GG" (default) or "IGIG".}

\item{tau2_0}{numeric. Value of global variance when \code{prior = "fixed"}.}
}
\value{
An object of class \code{\link{list}}
}
\description{
This function implements inference methods for the linear regression model with various global shrinkage priors.
}
\details{
This function fits the following model:

\deqn{y \mid X, \beta, \sigma^{2} \sim \text{N}_n(X\beta, \sigma^2 I_n)}
\deqn{\beta \mid \sigma^2,\tau^2 \sim \text{N}_p(0, \tau^2 \sigma^2 I_p)}
\deqn{p(\sigma^2) \propto \sigma^{-2}}

with the following choice of priors for \eqn{\tau^2}:

\itemize{
 \item{}{\eqn{\tau^2 \ \sim\  \text{InvGamma}(a, b)} when \code{prior = "invGamma"}}
 \item{}{\eqn{\tau^2 \ \sim\  \text{BetaPrime}(a, b)} when \code{prior = "BetaPrime"}}
 \item{}{\eqn{\tau^2 \ \sim\  \text{InvGaussian}(a, b)} when \code{prior = "invGaussian"}}
 \item{}{\eqn{\tau^2 \ \sim\  \text{Gamma}(a, b)} when \code{prior = "Gamma"}}
 \item{}{\eqn{\tau^2 \ =\  \hat{\tau}_{\text{ML}}^2} when \code{prior = "ml"}}
 \item{}{\eqn{\tau^2 \ =\  \tau_{0}^2} when \code{prior = "fixed"}}
} 

where \eqn{\text{Gamma}()} represents the Gamma distribution with
shape and rate parametrization and
\eqn{\hat{\tau}_{\text{ML}}^2} is the value of \eqn{\tau^2} that
maximizes the marginal likelihood (ML) of the model.
}
\examples{
\dontrun{
# Generate data
set.seed(3478)
n <- 200
p <- 100
X <- matrix(rnorm(n*p), n, p)
trueBeta <- rnorm(p, 0, 1)
y <- as.vector(X\%*\%trueBeta + rnorm(n, sd=1))

# Ridge with fixed \tau^2
set.seed(7216)
res0 <- brg(y, X, prior = "ml")

# Ridge with inverse-Gamma prior
set.seed(1525)
res1 <- brg(y, X, prior = "invGamma", a = 1e-05, b = 1e-05)

# Ridge with Beta-Prime prior
set.seed(5727)
res2 <- brg(y, X, prior = "BetaPrime", a = 0.5, b = 0.5)

# Ridge with inverse-Gaussian prior
set.seed(7804)
res3 <- brg(y, X, prior = "invGaussian", a = 1, b = 1e-05)

# Ridge with Gamma prior
set.seed(6570)
res4 <- brg(y, X, prior = "Gamma", a = 1e-05, b = 1e-05)

# Plot posterior densities of \tau^2
xs <- sapply(paste0("res", 1:4), function(x){get(x)$tau2s}, simplify=FALSE)
dens <- lapply(xs, density)
plot(NA, ylim=c(0, 1.8), xlim=c(0, 3), ylab="Posterior density", xlab=expression(tau^2))
#plot(NA, xlim=range(sapply(dens, "[", "x")), ylim=range(sapply(dens, "[", "y")))
mapply(lines, dens, col=1:length(dens), lwd=2)
labs <- c("invGamma", "BetaPrime", "invGaussian", "Gamma")
legend("topright", legend=labs, fill=1:length(dens), bty="n")
abline(v=res0$tau2s, lty=2, lwd=2)

}

}
\references{
Leday, G.G.R. et al. (2019)
}
\author{
Gwenael G.R. Leday
}
