% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brl.r
\name{brl}
\alias{brl}
\title{Bayesian linear regression with local shrinkage priors}
\usage{
brl(
  y,
  X,
  g = 1:ncol(X),
  prior = "BetaPrime",
  a = 0.5,
  b = 0.5,
  mcmc = 5000L,
  burnin = 1000L,
  thin = 10L,
  verbose = TRUE,
  output = "both",
  BP = "IGIG"
)
}
\arguments{
\item{y}{numeric response vector of length n.}

\item{X}{n by p data matrix.}

\item{g}{numeric vector of group memberships.}

\item{prior}{character. See Details.}

\item{a}{hyperparameter.}

\item{b}{hyperparameter.}

\item{mcmc}{integer. Number of desired samples.}

\item{burnin}{integer. Number of burn-in samples.}

\item{thin}{integer. Number of consecutive samples to skip.}

\item{verbose}{logical. Whether information on progress should be printed.}

\item{output}{character. Either "samples", "summary" or both.}

\item{BP}{character. Parametrization used when \code{prior = "BetaPrime"}. Either "GG" or "IGIG" (default).}
}
\value{
An object of class \code{\link{list}}
}
\description{
This function implements inference methods for the Bayesian regression model with local shrinkage priors.
}
\details{
This function fits the following model:

\deqn{y \mid X, \beta, \sigma^{2} \sim \text{N}_n(X\beta, \sigma^2 I_n)}
\deqn{\beta \mid \tau^2, \sigma^2 \sim \text{N}_p(0, D_{\tau} \sigma^{2})}
\deqn{p(\sigma^2) \propto \sigma^{-2}}

where \eqn{(D_{\tau})_{jj} = \tau_k^2} if variable \eqn{j = 1, \ldots, p}
belongs to group \eqn{k = 1, \ldots, K}. Group membership
is specified via the argument \code{g}.

The following choice of priors is available for \eqn{\tau_k^2}:

\itemize{
 \item{}{\eqn{\tau_k^2 \ \sim\  \text{InvGamma}(a, b)} when \code{prior = "invGamma"}}
 \item{}{\eqn{\tau_k^2 \ \sim\  \text{BetaPrime}(a, b)} when \code{prior = "BetaPrime"}}
 \item{}{\eqn{\tau_k^2 \ \sim\  \text{InvGaussian}(a, b)} when \code{prior = "invGaussian"}}
 \item{}{\eqn{\tau_k^2 \ \sim\  \text{Gamma}(a, b)} when \code{prior = "Gamma"}}
}
}
\examples{
\dontrun{
# Generate data
library(mvtnorm)
set.seed(20190719)
n <- 200
p <- 100
V <- toeplitz(c(0, 1, rep(0, p-2)))*0.25 + diag(p) # AR1(0.25)
X <- rmvnorm(n, mean = rep(0,p), sigma = V)
trueBeta <- c(1, -1, 0.9, -0.9, 0.8, -0.8, 0.7, -0.7, 0.6, -0.6, rep(0, p-10))
XB <- X\%*\%trueBeta
y <- as.vector(XB + rnorm(nrow(X), mean=0, sd=sqrt(.1)))
g <- rep(c(1, 2), p/2)

# Group ridge with inverse-Gamma prior
set.seed(2348)
res1 <- brl(y, X, g, prior = "invGamma", a = 1e-05, b = 1e-05)

# Group ridge with Beta-Prime prior
set.seed(7191)
res2 <- brl(y, X, g, prior = "BetaPrime", a = 0.5, b = 0.5)

# Group ridge with inverse-Gaussian prior
set.seed(6320)
res3 <- brl(y, X, g, prior = "invGaussian", a = 1, b = 1e-05)

# Group ridge with Gamma prior
set.seed(4230)
res4 <- brl(y, X, g, prior = "Gamma", a = 1e-05, b = 1e-05)

}

}
\references{
Leday, G.G.R. et al. (2019)
}
\author{
Gwenael G.R. Leday
}
