% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_df.R
\name{calculate_df}
\alias{calculate_df}
\title{Calculate the degrees of freedom for a model based on nonzero predictors.}
\usage{
calculate_df(x, betas)
}
\arguments{
\item{x}{A matrix of predictors.}

\item{betas}{A matrix of coefficients for each value of lambda.}
}
\value{
The degrees of freedom for the model.
}
\description{
This function takes a matrix of predictors (\code{x}) for all variables and a 
vector of coefficients (\code{betas}), and calculates the degrees of freedom, 
which is the rank of the covariate matrix for the selected variables.
}
\examples{
x <- matrix(rnorm(100), ncol = 5)
x <- cbind(x, x[,2]+x[,4])
betas <- matrix(c(0, 0.5, 0, 1, 0, 1))
rownames(betas) = sprintf("x\%d", 1:6)
calculate_df(x, betas)

}
