% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_selection.R
\name{coef.bestsubsetfit}
\alias{coef.bestsubsetfit}
\title{Extract coefficients from a bestsubsetfit object.}
\usage{
\method{coef}{bestsubsetfit}(object, ...)
}
\arguments{
\item{object}{Fitted \code{"alassofit"} model object}

\item{...}{functionless}
}
\description{
It returns shrunken regression coefficients if `shrinkage = TRUE` in the `bestsubsetfit` object
}
