% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimation.R
\name{coef.breiman}
\alias{coef.breiman}
\title{Extract coefficients from a breiman object}
\usage{
\method{coef}{breiman}(object, ...)
}
\arguments{
\item{object}{Fitted \code{"breiman"} object}

\item{...}{not useful at the moment}
}
\description{
Extract coefficients from a breiman object
}
