% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonnegative_garrote.R
\name{coef.garrote}
\alias{coef.garrote}
\title{Extract coefficients from a garrote object}
\usage{
\method{coef}{garrote}(object, ...)
}
\arguments{
\item{object}{Fitted \code{"garrote"} model}

\item{...}{Not used at the moment}
}
\value{
Returns the regression coefficients
}
\description{
Extract the coefficients from a fitted garrote model. The regression coefficients
are provided without the intercept.
}
