% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.R
\name{coef.lassofit}
\alias{coef.lassofit}
\title{Extract coefficients from a lassofit object}
\usage{
\method{coef}{lassofit}(object, ...)
}
\arguments{
\item{object}{Fitted \code{"lassofit"} model object}

\item{...}{not used at the moment}
}
\value{
regression estimates without intercept
}
\description{
Extract coefficients from a lassofit object
}
