% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classical_linear_models.R
\name{coef.linearmodel}
\alias{coef.linearmodel}
\title{Extract coefficients from a "linearmodel" object}
\usage{
\method{coef}{linearmodel}(object, ...)
}
\arguments{
\item{object}{Fitted \code{"linearmodel"} model object}

\item{...}{not used at the moment}
}
\description{
Extract coefficients from a "linearmodel" object
}
