% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonnegative_garrote.R
\name{coef.nngfit}
\alias{coef.nngfit}
\title{Extract coefficients from nngfit object}
\usage{
\method{coef}{nngfit}(object, ...)
}
\arguments{
\item{object}{Fitted \code{"nngfit"} object}

\item{...}{not used at the moment}
}
\description{
Extract coefficients from nngfit object
}
