% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimation.R
\name{coef.pwsf}
\alias{coef.pwsf}
\title{Extract shrunken regression estimates from pwsf object}
\usage{
\method{coef}{pwsf}(object, ...)
}
\arguments{
\item{object}{Fitted \code{"pwsf"} object}

\item{...}{not used at the moment}
}
\description{
Extract shrunken regression estimates from pwsf object
}
