% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{coef.ridgefit}
\alias{coef.ridgefit}
\title{Extract coefficients from a ridgefit object}
\usage{
\method{coef}{ridgefit}(object, ...)
}
\arguments{
\item{object}{Fitted \code{"ridgefit"} model object}

\item{...}{not used at the moment}
}
\value{
regression estimates without intercept
}
\description{
Extract coefficients from a ridgefit object
}
