% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_simulations.R
\name{combine_sim_coef}
\alias{combine_sim_coef}
\title{Combines regression coefficients or shrinkage factors}
\usage{
combine_sim_coef(
  path,
  type = c("betas", "shrinkage"),
  add_rep = FALSE,
  freq = FALSE
)
}
\arguments{
\item{path}{The directory where the RDS files for each simulation scenario
are stored. Example: "C:/Users/Desktop/varselection/"}

\item{type}{Specifies whether to return "betas" or "shrinkage". Default is "betas",
which returns a dataframe of regression estimates for each method and
replication of a scenario.}

\item{add_rep}{Specifies whether to include simulation replication in the output. Important for
calculating similarities. Set it to TRUE if `compute_similarity_sim()` function will be used.}

\item{freq}{A logical value that indicates whether the regression estimates of
variables should be converted to 0s and 1s. This is useful if the inclusion
frequency of variables is required. If \code{freq = TRUE} and
\code{plot_coefs()} is used, the inclusion frequencies will be plotted.}
}
\value{
A dataframe for each scenario without aggregation. This means that
each scenario is repeated for each replication. If the number of replications
is 5, then each scenario will be replicated 5 times for each regression method.
}
\description{
#The function combines the output stored in an object of class sim_master
(produced by \code{\link{sim_master}}) into a dataframe. In cases where
different numbers of variables are used, the function combines the inputs
into one large data frame. If there are missing columns in the smaller data
frames, the function fills those columns with NAs in the combined data frame.
}
