% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_simulations.R
\name{combine_sim_metrics}
\alias{combine_sim_metrics}
\title{Combine simulation output}
\usage{
combine_sim_metrics(
  path,
  metric = c("rr", "rte", "trainerror", "valerror", "testerror", "nullerror", "prop",
    "risk", "nullrisk", "nonzero", "fpos", "tpos", "fneg", "tneg", "F1", "MCC", "fpr",
    "fnr", "errorrate", "optimism", "lambda", "gamma", "runtime"),
  add_rep = FALSE
)
}
\arguments{
\item{path}{the directory where the rds files are stored for each simulation
scenario. E.g "C:/Users/ek/Desktop/varselection/"}

\item{metric}{the metric of interest. Available metrics include:}

\item{add_rep}{Specifies whether to include simulation replication in the output. Important for
 calculating similarities. Set it to TRUE if `compute_similarity_sim()` function will be used.
\itemize{
  \item `rr` - relative risk which measures the risk or model error relative to the null model (model without predictors).
  \item `rte` - relative test error which measures the expected test error relative to the Bayes error rate.
  \item `trainerror` - error on the training data.
  \item `valerror` - Error on the validation data.
  \item `testerror` - error on the test data.
  \item `nullerror` - error for a model without predictors.
  \item `prop` - proportion of variance unexplained by the model.
  \item `risk` - model error.
  \item `nullrisk` - error in the null model.
  \item `nonzero` - Number of Nonzero Coefficients.
  \item `fpos` - false Positives.
  \item `tpos` - true Positives.
  \item `fneg` - false Negatives.
  \item `tneg` - True Negatives.
  \item `F1` - F-score which measures the accuracy of the support recovery.
  \item `MCC` - Matthews correlation coefficient which measures the accuracy of the support recovery.
  \item `fpr` - False Positive Rate which measures the proportion of actual negatives that are predicted as positives.
  \item `fnr` - False Negative Rate which measures the proportion of actual positives that are predicted as negatives.
  \item `errorrate` - Error rate.
  \item `optimism` - Optimism of training error is computed as (test error - train error)/train error.
  \item `runtime` - Time used in simulation runs.
  \item `lambda` - Tuning Parameter for Regression Models.
  \item `gamma` - Tuning Parameter for Regression Model.
}}
}
\value{
A dataframe containing the combined simulation output for the specified metric.
}
\description{
Combine simulation output stored in an object of class 'sim_master' (produced
by \code{\link{sim_master}}) into a dataframe for further processing. This
function is not suitable for regression estimates or shrinkage factors.
Please refer to the 'combine_sim_coef' function for those cases.
}
