% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data_master.R
\name{corr_type}
\alias{corr_type}
\title{Generates correlation structure}
\usage{
corr_type(
  nvar = 15,
  type = c("c1", "c2", "c3", "c4", "c5"),
  withinSignalCorr = 0.8,
  withinNoiseCorr = 0.8,
  betweenSignalNoiseCorr = 0,
  betatype = c("a", "b", "c", "d")
)
}
\arguments{
\item{nvar}{The number of variables. Default is 15}

\item{type}{correlation types. `c1` = Houwelingen and Sauerbrei (2013),
`c2` = AR1(0.3), `c3` = AR1(0.8), `c4` = Educational Body fat and `c5` = block
correlation matrix.}

\item{withinSignalCorr, withinNoiseCorr, betweenSignalNoiseCorr}{Specifies within
signal correlation,within noise correlation and between signal and noise 
correlation, respectively. Only applicable if `type = "c5"`.}

\item{betatype}{Specifies the true regression coefficients. See `beta_type()` function for
details.}
}
\description{
Generates four types of correlation structure as described in Kipruto and Sauerbrei (2022).
In addition, it generates block correlation matrix. Block correlation matrix uses the
true regression coefficients to determine the indices of signal and noise variables 
to create the two groups.
}
\references{
Kipruto, E. and Sauerbrei, W. (2022). Comparison of variable 
selection procedures and investigation of the role of shrinkage in linear
regression-protocol of a simulation study in low-dimensional data. Plos one, 
17(10), p.e0271240.
}
