% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonnegative_garrote.R
\name{cv.garrote}
\alias{cv.garrote}
\title{Cross-validation for nonnegative garrote}
\usage{
cv.garrote(
  x,
  y,
  nfolds = 10,
  foldid = NULL,
  lambda = NULL,
  alpha = 1,
  nlambda = 50,
  lower.limits = 0,
  upper.limits = Inf,
  standardize = FALSE,
  standardize.response = FALSE,
  lambda.initial = NULL,
  beta.initial = NULL,
  initial.estimates = c("default", "ridge", "lasso"),
  type.measure = c("mse", "mae"),
  lambda.min.ratio = 1e-04
)
}
\arguments{
\item{x}{A standardized matrix of predictors of dimension nobs x nvars, where
each row is an observation vector.If not standardized, set standardize to TRUE.}

\item{y}{A numeric centered quantitative vector of response. If not centered,
set `standardize` to TRUE; this will standardize x but center y. See `standardize.response`
for more details.}

\item{nfolds}{The number of folds used when cross-validating lambda as well
as the lasso and ridge initial estimates. Default is 10.}

\item{foldid}{an optional vector of values between 1 and nfold identifying
what fold each observation is in. If supplied, nfolds can be missing.}

\item{lambda}{Optional user-supplied sequence of lambdas. The default is NULL, and
the program chooses its own sequence. If lambda is supplied, its length must be at least two.}

\item{alpha}{The mixing parameter where alpha = 1 represents the usual nonnegative
garrote, while alpha = 0 represents the garrote with a ridge-like penalty.}

\item{nlambda}{An integer specifying the number of values of lambda
for both nonnegative garrote and initial estimators (ridge and lasso). The default is NULL,
which sets `nlambda` to 100.}

\item{lower.limits}{Vector of lower limits for each shrinkage factor.
The default is 0, which will be replicated to have a length equal to the number of predictors, and
nonnegative shrinkage factors will be estimated.}

\item{upper.limits}{Vector of upper limits for the shrinkage factors. The default is Inf, and
unbounded shrinkage factors will be estimated.}

\item{standardize}{Specifies whether to standardize the predictors matrix
\code{`x`} to have mean 0 and unit variance. If set to TRUE, \code{`x`} will be
standardized. Additionally, the response variable \code{`y`} will be centered
by subtracting its mean. This standardization step can be useful to ensure
that the predictors are on a comparable scale. By default,
\code{standardize = FALSE}, indicating that no standardization will be performed.
This assumes that users have already standardized x and centered y so that the intercept
is zero.}

\item{standardize.response}{Specifies whether the response variable y should be
standardized to have unit variance. This is important if standardize.response was
set to TRUE in `cv.garrote()`. This option divides \code{`y`} by its standard deviation.}

\item{lambda.initial}{An optional tuning parameter for the initial estimates
using lasso or ridge regularization. If a vector of lambda is provided,the
optimal will be determined by CV}

\item{beta.initial}{A vector of initial estimates obtained from standardized
predictor matrix x. The default value is \code{beta.initial = NULL}, and the
program will estimate them based on the selected `initial.estimate` and
`lambda.initial`.}

\item{initial.estimates}{Specifies the type of initial estimates to use for
constructing adaptive lasso weights. The \code{initial.estimates = "default"}
option uses \code{"OLS"} estimates. Other options include
\code{initial.estimates ="ridge"} or \code{initial.estimates ="lasso"}
estimates, which are calculated using the user-supplied \code{`lambda.initial`.}}

\item{type.measure}{Loss function to use for cross-validation for both nonnegative garrote and
initial estimates (ridge or lasso) when not provided. Currently, two options are available. The
default option is \code{type.measure = "mse"}, which corresponds to squared error. Another
option is \code{type.measure = "mae"}, which corresponds to mean absolute error.}

\item{lambda.min.ratio}{Lambda can be provided if the user wants to specify
the lambda sequence, but the typical usage is for the program to construct the
lambda sequence on its own. When automatically generated, the lambda sequence
is determined by the ratio of lambda.max and lambda.min. The program generates nlambda values
on the log scale from lambda.max down to lambda.min. Lambda.max is not user-specified but is
computed from the input standardized x and y. It is the smallest value for lambda in which all
the estimated coefficients are zero. The default is `lambda.min.ratio = 0.0001`.}
}
\value{
Returns \code{cv.glmnet} objects. Please refer to the return
section of the mentioned function in the \code{glmnet} package. Additionally,
it returns the lambda (lambda.initial) for the optimal lasso or ridge initial
estimates.
}
\description{
Perform k-fold cross-validation for nonnegative garrote and return the optimal value of
lambda. The estimated value of lambda depends on the initial estimate used.
}
