% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_df.R
\name{df_ridge}
\alias{df_ridge}
\title{Compute the degrees of freedom for ridge regression.}
\usage{
df_ridge(x, lambda)
}
\arguments{
\item{x}{A numeric matrix representing the design matrix.}

\item{lambda}{A numeric vector representing the sequence of lambda values.}
}
\value{
A numeric vector containing the degrees of freedom for each lambda value.
}
\description{
This function calculates the degrees of freedom for ridge regression for a given design matrix and a sequence of lambda values.
}
\examples{
# Example usage:
x <- matrix(rnorm(20), ncol = 2)
lambda_values <- c(0.1, 0.01, 0.001)
df_ridge_result <- df_ridge(x, lambda_values)

}
\references{
see element of statistical learning for the formula. eqn 3.5
}
