% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eightnum.r
\name{eightnum}
\alias{eightnum}
\title{eight-number summary}
\usage{
eightnum(x)
}
\arguments{
\item{x}{numeric vector of samples}
}
\value{
A numeric vector
}
\description{
This function computes an eight-number summary for a sample
}
\details{
This function returns the mean, sd, min, median, max, quantiles and mode of the sample.
}
\examples{
\dontrun{
set.seed(12345)
xx <- rnorm(10000)
eightnum(xx)
}

}
\references{
Leday, G.G.R. et al. (2019)
}
\author{
Gwenael G.R. Leday
}
