% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_simulations.R
\name{extract_sim_scenario}
\alias{extract_sim_scenario}
\title{Extract Simulation Scenario Data from RDS Files}
\usage{
extract_sim_scenario(files_paths)
}
\arguments{
\item{files_paths}{A character vector of file paths to the RDS files.}
}
\value{
A data frame containing the extracted information from all the RDS files.
Each row corresponds to a simulation scenario and includes the following columns:
\describe{
  \item{n}{The sample size of the simulation scenario.}
  \item{snr}{The signal-to-noise ratio of the simulation scenario.}
  \item{betatype}{The type of beta used in the simulation scenario.}
  \item{corrtype}{The type of correlation used in the simulation scenario.}
  \item{nvar}{The number of variables used in the simulation scenario.}
}
}
\description{
This function reads multiple RDS files containing simulation scenario data,
extracts relevant information, and combines it into a single data frame.
}
\examples{
\dontrun{
# Assuming you have RDS files are stored in the path
path <- "Q:/x/postestimation/simdata/final/"
sim_data <- extract_sim_scenario(paths)
print(sim_data)
}
}
