% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonnegative_garrote.R
\name{garrote}
\alias{garrote}
\title{Fit a Gaussian linear model with nonnegative garrote regularization}
\usage{
garrote(
  x,
  y,
  lambda,
  lambda.initial = NULL,
  beta.initial = NULL,
  alpha = 1,
  standardize = FALSE,
  standardize.response = FALSE,
  lower.limits = 0,
  upper.limits = Inf,
  initial.estimates = c("default", "ridge", "lasso")
)
}
\arguments{
\item{x}{A standardized matrix of predictors of dimension nobs x nvars, where
each row is an observation vector.If not standardized, set standardize to TRUE.}

\item{y}{A numeric centered quantitative vector of response. If not centered,
set `standardize` to TRUE; this will standardize x but center y. See `standardize.response`
for more details.}

\item{lambda}{Regularization parameter that controls the strength of the penalty on shrinkage
factors.}

\item{lambda.initial}{The tuning parameter for the initial estimates using
lasso or ridge regularization. If \code{initial.estimates = "default"},
`lambda.initial` can be NULL since OLS estimates do not require a tuning
parameter. However, if \code{initial.estimates = "ridge"} or
\code{initial.estimates = "lasso"}, the tuning parameters must be supplied.}

\item{beta.initial}{A vector of initial estimates obtained from standardized
predictor matrix x. The default value is \code{beta.initial = NULL}, and the
program will estimate them based on the selected `initial.estimate` and
`lambda.initial`.}

\item{alpha}{The mixing parameter where alpha = 1 represents the usual nonnegative
garrote, while alpha = 0 represents the garrote with a ridge-like penalty.}

\item{standardize}{Specifies whether to standardize the predictors matrix
\code{`x`} to have mean 0 and unit variance. If set to TRUE, \code{`x`} will be
standardized. Additionally, the response variable \code{`y`} will be centered
by subtracting its mean. This standardization step can be useful to ensure
that the predictors are on a comparable scale. By default,
\code{standardize = FALSE}, indicating that no standardization will be performed.
This assumes that users have already standardized x and centered y so that the intercept
is zero.}

\item{standardize.response}{Specifies whether the response variable y should be
standardized to have unit variance. This is important if standardize.response was
set to TRUE in `cv.garrote()`. This option divides \code{`y`} by its standard deviation.}

\item{lower.limits}{Vector of lower limits for each shrinkage factor.
The default is 0, which will be replicated to have a length equal to the number of predictors, and
nonnegative shrinkage factors will be estimated.}

\item{upper.limits}{Vector of upper limits for the shrinkage factors. The default is Inf, and
unbounded shrinkage factors will be estimated.}

\item{initial.estimates}{Specifies the type of initial estimates to use for
constructing nonnegative garrote weights. The \code{initial.estimates = "default"}
option uses \code{"OLS"} estimates. Other options include
\code{initial.estimates ="ridge"} or \code{initial.estimates ="lasso"}
estimates, which are calculated using the user-supplied \code{`lambda.initial`.}}
}
\value{
Returns the following items:
 \item{beta}{Shrunken regression estimates without an intercept}
 \item{shrinkageFactors}{Shrinkage factors for all regression estimates}
 \item{lambda}{Lambda value used for tuning the model}
 \item{x}{Matrix \code{x} used to fit the model}
 \item{y}{Values of \code{y} used to fit the model}
}
\description{
Fits the nonnegative garrote regression for continuous outcomes proposed by
Breiman (1995). Currently, it does not support other types of outcomes such as categorical or
survival.
}
