% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimation.R
\name{global}
\alias{global}
\title{Estimate global shrinkage factors}
\usage{
global(
  x,
  y,
  varx = NULL,
  nfolds = 10,
  foldid = NULL,
  standardize = FALSE,
  nonnegative = FALSE
)
}
\arguments{
\item{x}{A standardized matrix of predictors with dimensions nobs x nvars, where 
each row represents an observation vector. If not standardized, set standardize to
TRUE.}

\item{y}{A numeric, centered quantitative vector of responses. If not centered,
set `standardize` to TRUE; this will standardize x but center y.}

\item{varx}{Names of the variables selected by a procedure. Used to subset 
the predictor matrix \code{`x`}. The default value is NULL, and all column
names of x are used.}

\item{nfolds}{Number of cross-validation folds. The default is 10; setting 
`nfolds = n` corresponds to leave-one-out cross-validation.}

\item{foldid}{An optional vector of values between 1 and \code{"nfolds"} that 
identifies the fold to which each observation belongs. The default is 
NULL, and the program will generate it.}

\item{standardize}{Specifies whether to standardize the predictors matrix 
\code{`x`} to have mean 0 and unit variance. If set to TRUE, \code{`x`} will be
standardized. Additionally, the response variable \code{`y`} will be centered
by subtracting its mean. This standardization step can be useful to ensure 
that the predictors are on a comparable scale. By default, 
\code{standardize=FALSE}, indicating that no standardization will be performed.
This assumes that users have already standardized their data.}

\item{nonnegative}{Specify whether nonnegative shrinkage factors should be 
estimated. The default is FALSE, allowing for the estimation of negative 
shrinkage factors. Negative shrinkage factors are rarely estimated unless 
in extreme situations where Signal-to-noise ratio is extremely low.}
}
\value{
Returns a list with the following components:
\item{ShrunkenRegCoef}{Shrunken regression estimates for the selected variables,
with coefficients of zero assigned to the unselected variables.}
\item{ShrinkageFactors}{Shrinkage factor for the selected variables, with 
shrinkage factors of zero assigned to the unselected variables.}
\item{variables_selected}{The variables chosen for estimating shrinkage
factors (varx)}
\item{nvar}{The number of variables selected by the procedure (length of 
\code{"varx"}).}
}
\description{
Estimates global shrinkage factors and calculates corresponding 
shrunken regression estimates using either k-fold cross-validation (CV) or 
leave-one-out CV. Global shrinkage involves shrinking all regression coefficients
by the same factor, as proposed by van Houwelingen and le Cessie (1990) and 
Verweij and van Houwelingen (1993).
}
