% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridge.r
\name{gridge}
\alias{gridge}
\title{Bayesian group ridge regression}
\usage{
gridge(y, X, g, prior = "Gamma", c = NULL, mcmc = 5000L,
  burnin = 1000L, thin = 10L, eb = 1000L, verbose = TRUE,
  light = FALSE)
}
\arguments{
\item{y}{response vector of length n.}

\item{X}{n by p data matrix.}

\item{g}{vector of length p for group memberships.}

\item{prior}{character. Either "Gamma" or "BetaPrime". See Details.}

\item{c}{hyperparameter.}

\item{mcmc}{integer. Number of desired samples.}

\item{burnin}{integer. Number of burn-in samples.}

\item{thin}{integer. Number of consecutive samples to skip.}

\item{eb}{integer. Empirical Bayes carried out every 'eb' iteration.}

\item{verbose}{logical. Whether information on progress should be be printed.}

\item{light}{logical. If TRUE, only return a summary of the samples.}
}
\value{
An object of class \code{\link{list}}
}
\description{
This function implements a Gibbs algorithm for the ridge regression model with
grouped variables (predictors).
}
\details{
blablabla
}
\examples{
\dontrun{
# Generate data
library(mvtnorm)
set.seed(20190719)
n <- 200
p <- 100
V <- toeplitz(c(0, 1, rep(0, p-2)))*0.25 + diag(p) # AR1(0.25)
X <- rmvnorm(n, mean = rep(0,p), sigma = V)
trueBeta <- c(1, -1, 0.9, -0.9, 0.8, -0.8, 0.7, -0.7, 0.6, -0.6, rep(0, 90))
XB <- X\%*\%trueBeta
y <- as.vector(XB + rnorm(nrow(X), mean=0, sd=sqrt(.1)))
g <- rep(c(1, 2), 50)

# Run gibbs sampler
res <- gridge(y, X, g)

# Extract summary
res$summary
}

}
\references{
Leday, G.G.R. et al. (2019)
}
\author{
Gwenael G.R. Leday
}
