% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonnegative_garrote.R
\name{lambda_seq_alasso}
\alias{lambda_seq_alasso}
\title{Calculates a sequence of lambdas for adaptive lasso regularization.}
\usage{
lambda_seq_alasso(
  x,
  y,
  beta.initial,
  nlambda = 100,
  epsilon = 1e-04,
  gamma = NULL,
  ignore.gamma = FALSE
)
}
\arguments{
\item{x}{A standardized matrix of predictors}

\item{y}{A standardized vector of response}

\item{beta.initial}{Initial estimates for the standardized matrix x.}

\item{nlambda}{Number of lambdas}

\item{epsilon}{The ratio of smallest value of the generated lambda
sequence (say lambda.min) to lambda.max. See lambda.min.ratio
in \code{[cv.glmnet]} for more details}

\item{gamma}{The tuning parameter for adaptive lasso weights. The default
value is NULL, and the program uses seq(0.5, 2, 0.5).}

\item{ignore.gamma}{Specifies whether to ignore values of gamma supplied by the
user and use default values seq(0.5, 2, 0.5)}
}
\value{
Returns sequence of lambdas
}
\description{
Calculates the sequence of lambdas of length `nlambda`. If gamma
is a vector, it calculates all the lambdas for each gamma, then finds the
overall maximum and minimum, and uses them to compute the final sequence
of lambdas on a logarithmic scale. It assumes that x and y are standardized.
}
