% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonnegative_garrote.R
\name{lambda_seq_garrote}
\alias{lambda_seq_garrote}
\title{Calculate the values of lambdas for the nonnegative garrote}
\usage{
lambda_seq_garrote(
  x,
  y,
  beta.initial,
  nlambda = 100,
  epsilon = 1e-04,
  alpha = 1
)
}
\arguments{
\item{x}{A standardized matrix of predictors.}

\item{y}{A standardized vector of the response variable.}

\item{beta.initial}{Initial estimates for the standardized matrix x.}

\item{nlambda}{The number of lambdas to be returned.}

\item{epsilon}{The smallest value used in calculating the minimum lambda. The default value is
0.0001, referred to elsewhere as `lambda.min.ratio`.}

\item{alpha}{The mixing parameter, where alpha = 1 represents the usual nonnegative
garrote, while alpha = 0 represents the garrote with a ridge-like penalty.}
}
\value{
Returns a sequence of lambdas with a length of nlambda.
}
\description{
This function calculates the sequence of lambdas with a specified length, nlambda. It
assumes that both x and y are already standardized.
}
