% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.R
\name{lambda_seq_lasso}
\alias{lambda_seq_lasso}
\title{calculates sequence of lambdas for the lasso}
\usage{
lambda_seq_lasso(x, y, nlambda = 100, epsilon = 1e-04)
}
\arguments{
\item{x}{a standardized matrix of predictors}

\item{y}{a centered vector of response}

\item{nlambda}{Number of lambdas}

\item{epsilon}{The ratio of smallest value of the generated lambda
sequence (say lambda.min) to lambda.max. See lambda.min.ratio
in \code{[cv.glmnet]} for more details}
}
\description{
calculates sequence of lambdas for the lasso
}
