% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{lambda_seq_ridge}
\alias{lambda_seq_ridge}
\title{Function that calculates lambda sequence for the ridge}
\usage{
lambda_seq_ridge(x, y, nlambda = 100, epsilon = 1e-04)
}
\arguments{
\item{x}{a standardized matrix of predictors}

\item{y}{a centered vector of response}

\item{nlambda}{Number of lambdas}

\item{epsilon}{The ratio of smallest value of the generated lambda
sequence (say lambda.min) to lambda.max. See lambda.min.ratio
in \code{[cv.glmnet]} for more details}
}
\description{
Function that calculates lambda sequence for the ridge
}
