% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classical_linear_models.R
\name{linear_model}
\alias{linear_model}
\title{Fit the Gaussian linear regression model with and without shrinkage.}
\usage{
linear_model(
  x,
  y,
  foldid = NULL,
  nfolds = 10,
  shrinkage = c("none", "pwsf", "global", "breiman"),
  choice = c("tenfold", "loocv"),
  lower.limits = -Inf,
  upper.limits = Inf,
  type.measure = c("mse", "mae"),
  nonnegative = FALSE,
  standardize = FALSE,
  sigma2 = NULL,
  betatypes = NULL
)
}
\arguments{
\item{x}{A standardized matrix of predictors with dimensions nobs x nvars, where 
each row represents an observation vector. If not standardized, set standardize to
TRUE.}

\item{y}{A numeric, centered quantitative vector of responses. If not centered,
set `standardize` to TRUE; this will standardize x but center y.}

\item{foldid}{A vector of values between 1 and nfolds identifying what fold
each observation is in while conducting cross-validation. Default is NULL
and the program will generate it.}

\item{nfolds}{Number of folds for cross-validation. Default is 10. If k = n then k-fold is
equal to leave-one-out cross-validation.}

\item{shrinkage}{Specifies the method for shrinking parameter estimates.
Three options are available: Parameter-wise shrinkage (PWS), global shrinkage, and 
quadratic PWS (Breiman method). The default is `none`, and no shrinkage is conducted.}

\item{choice}{Specifies which method of estimating shrinkage factors to use: 
tenfold cross-validation or leave-one-out cross-validation. Default is 
tenfold cross-validation.}

\item{lower.limits}{Vector of length nvar that specifies the lower limits for the shrinkage
factor when using Breiman's method. The default is -Inf, potentially allowing estimation of 
negative shrinkage factors. If \code{nonnegative = TRUE}, the program will set 
\code{lower.limits = rep(0, nvar)} for variables, leading to the estimation of nonnegative 
shrinkage factors. In short, setting  \code{lower.limits = 0} is identical to setting 
\code{nonnegative = TRUE}. This option is only applicable to the Breiman method.}

\item{upper.limits}{Vector of length nvar specifying upper limits for the shrinkage factors. 
The default is Inf, and unbounded shrinkage factors will be estimated. This 
option is only applicable to the Breiman method.}

\item{type.measure}{Loss to use for cross-validation. Currently, two options
are available. The default is `type.measure = "mse"`, which uses squared error. 
Another option is `type.measure = "mae"` (mean absolute error). This
option is only applicable to the Breiman method.}

\item{nonnegative}{Specify whether nonnegative shrinkage factors should be 
estimated. The default is FALSE, which means that negative shrinkage factors may be
 estimated. See the "lower.limits" argument for alternative instructions on estimating
  nonnegative shrinkage factors for the Breiman method.}

\item{standardize}{Specifies whether to standardize the predictors matrix 
\code{`x`} to have mean 0 and unit variance. If set to TRUE, \code{`x`} will be
 standardized. Additionally, the response variable \code{`y`} will be centered
 by subtracting its mean. By default,  \code{standardize = FALSE}, indicating that no 
 standardization will be performed. This assumes that users have already standardized x 
 and centered y so that the intercept is zero.}

\item{sigma2}{Residual variance obtained by fitting the full model(model without variable
selection). This value is utilized in the calculation of AIC and BIC, although it is 
currently not in use.}

\item{betatypes}{Not used but added for consistency with the oracle_model().}
}
\value{
: Returns a list with the following components:
\item{beta:}{Regression estimates. If shrinkage = "none", unshrunken OLS 
estimates are returned; otherwise, shrunken regression estimates are returned.}
\item{shrinkageFactors:}{Shrinkage factors for each predictor estimated 
using one of the specified shrinkage methods. If shrinkage = "none", the
shrinkage factors of 1s are returned.}
\item{nvar:}{The number of variables in the model.}
\item{x:}{A standardized matrix of predictors used in fitting the linear model.}
\item{y:}{A centered vector of the response variable used in fitting the 
linear model.}
\item{lambda,gamma:}{Parameters used in penalized methods, such as adaptive 
lasso. It is not applicable in the current context. However, the function 
returns the two parameters for consistency with the interface of penalized methods.}
}
\description{
Fits a linear regression model for a Gaussian response. The regression 
estimates can be subjected to shrinkage obtained from three different approaches: global
 shrinkage, parameter-wise shrinkage (PWS), and quadratic PWS (Breiman method).
}
