% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik.r
\name{log_lik}
\alias{log_lik}
\title{Obtain samples from the log-likelihood}
\usage{
log_lik(object, y, X, output = "both")
}
\arguments{
\item{object}{list object returned by brg(), brl(), brgl()...}

\item{y}{response vector.}

\item{X}{data matrix.}

\item{output}{character. Either "samples", "summary" or "both".}
}
\value{
An object of class \code{\link{list}}
}
\description{
This function provides mcmc samples of the log-likelihood for each observation.
}
\references{
Leday, G.G.R. et al. (2019)
}
\author{
Gwenael G.R. Leday
}
