% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.r, R/log_likelihood.r
\name{likelihood}
\alias{likelihood}
\alias{log_likelihood}
\title{Obtain samples from the (log-)likelihood}
\usage{
likelihood(object, y, X, output = "both")

log_likelihood(object, y, X, output = "both")
}
\arguments{
\item{object}{list object returned by brg() or brl()}

\item{y}{response vector.}

\item{X}{data matrix.}

\item{output}{character. Either "samples", "summary" or "both".}
}
\value{
An object of class \code{\link{list}}

An object of class \code{\link{list}}
}
\description{
This function provides samples and summaries of the (log-)likelihood for each observation.

This function provides samples and summaries of the (log-)likelihood for each observation.
}
\references{
Leday, G.G.R. et al. (2019)

Leday, G.G.R. et al. (2019)
}
\author{
Gwenael G.R. Leday
}
