% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_selection.R
\name{logl}
\alias{logl}
\title{estimates loglikehood of a Gaussian model}
\usage{
logl(res, w = NULL)
}
\arguments{
\item{res}{residuals from a fitted model. Should not contain missing values}

\item{w}{an optional vector of weights used in the fitting process. Should
be NULL or a numeric vector. see `[lm()]` for more details}
}
\description{
estimates loglikehood of a Gaussian model
}
