% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonnegative_garrote.R
\name{nng_aic_bic}
\alias{nng_aic_bic}
\title{Estimates the value of lambda for NNG based on AIC and BIC}
\usage{
nng_aic_bic(
  x,
  y,
  beta.initial,
  lambda.initial,
  criterion,
  initial.estimates,
  lower.limits,
  upper.limits,
  lambda,
  alpha,
  sigma2
)
}
\arguments{
\item{x}{A standardized matrix of predictors of dimension nobs x nvars, where
each row is an observation vector.If not standardized, set standardize to TRUE.}

\item{y}{A numeric centered quantitative vector of response. If not centered,
set `standardize` to TRUE; this will standardize x but center y. See `standardize.response`
for more details.}

\item{beta.initial}{A vector of initial estimates obtained from standardized
predictor matrix x. The default value is \code{beta.initial = NULL}, and the
program will estimate them based on the selected `initial.estimate` and
`lambda.initial`.}

\item{lambda.initial}{The tuning parameter for the initial estimates using
lasso or ridge regularization. If \code{initial.estimates = "default"},
`lambda.initial` can be NULL since OLS estimates do not require a tuning
parameter. However, if \code{initial.estimates = "ridge"} or
\code{initial.estimates = "lasso"}, the tuning parameters must be supplied.}

\item{criterion}{The criterion used to select tuning parameters for the
nonnegative garrote. Available options include: \code{criterion = "cv"}
which perform cross-validation to select the optimal tuning parameter, while
\code{criterion = "aic"} and \code{criterion = "bic"} select the tuning parameter
 based on the AIC and BIC, respectively.}

\item{initial.estimates}{Specifies the type of initial estimates to use for
constructing nonnegative garrote weights. The \code{initial.estimates = "default"}
option uses \code{"OLS"} estimates. Other options include
\code{initial.estimates ="ridge"} or \code{initial.estimates ="lasso"}
estimates, which are calculated using the user-supplied \code{`lambda.initial`.}}

\item{lower.limits}{Vector of lower limits for each shrinkage factor.
The default is 0, which will be replicated to have a length equal to the number of predictors, and
nonnegative shrinkage factors will be estimated.}

\item{upper.limits}{Vector of upper limits for the shrinkage factors. The default is Inf, and
unbounded shrinkage factors will be estimated.}

\item{lambda}{Optional user-supplied sequence of lambdas. The default is NULL, and
the program chooses its own sequence. If lambda is supplied, its length must be at least two.}

\item{alpha}{The mixing parameter where alpha = 1 represents the usual nonnegative
garrote, while alpha = 0 represents the garrote with a ridge-like penalty.}

\item{sigma2}{The residual variance obtained by fitting the full OLS model without
any regularization or feature selection. It is used for the calculation of AIC and BIC.
To compute these metrics, the residual variance needs to be provided via the
`sigma2` parameter.}
}
\description{
The value of lambda that yields the smallest AIC or BIC is considered the best lambda
for tuning the nonnegative garrote.
}
