% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_coefs}
\alias{plot_coefs}
\title{Plots regression estimates, shrinkage factors or inclusion frequencies}
\usage{
plot_coefs(
  x,
  betatypes = c("a", "b", "c", "d"),
  corrtypes = c("c1", "c2", "c3", "c4", "c5"),
  samplesize = NULL,
  snrx = NULL,
  nvar = 15,
  methods = NULL,
  plotype = c("means", "distribution"),
  legend.position = c("right", "bottom", "top", "left", "none"),
  row = c("betatype", "corrtype"),
  col = c("corrtype", "betatype"),
  varx = NULL,
  truncate.lower = NULL,
  truncate.upper = NULL,
  remove.zeros = FALSE,
  linetype = FALSE,
  add.se = FALSE,
  lwd = 0.7,
  points_shape = 16,
  points_size = 2,
  main = NULL,
  ylim = NULL,
  boxwidth = 0.5,
  facet = FALSE
)
}
\arguments{
\item{x}{The output of the \code{combine_sim_coef()} function. If inclusion
frequencies are required, the user must set \code{freq = TRUE} in
\code{combine_sim_coef()} and \code{plotype = "means"} in this function.}

\item{betatypes}{the beta type used in the simulation. Default is `betatypes = "a"`}

\item{corrtypes}{the correlation type type used in the simulation.
Default is `corrtypes = "a"`.}

\item{samplesize}{a vector of sample size used in the simulation. Default is NULL
and all the sample sizes used are plotted.}

\item{snrx}{a vector of signal to noise ratio used in the simulation. Default is NULL
and all the snr used are plotted.}

\item{nvar}{a numeric value for the number of variables used in the simulation.
Default is 15 variables}

\item{methods}{the name(s) of the method used in the simulation. Default is
null and all methods are plotted.}

\item{plotype}{the plot to generate. Default is `plotype = "means"` which plots
the average values of a given metric. If distributions are required,
set `plotype = "distribution"` to generate boxplots.}

\item{legend.position}{the position of the legend in the plot. Default is to
place it on the right side of the plot}

\item{row, col}{One of `betatype` or `corrtype`,indicating the variables for
the rows and columns for plotting facet grid; note that row and col must be
different. Default is `row = betatype` and `col = corrtype`, so that the
plotting grid displays the given metric versus the SNR or sample size,
in a plotting grid with the coefficient types across the rows, and the
correlation levels across the columns.}

\item{varx}{The names of variables to plot. The default value is NULL,
indicating that all variables will be plotted by default. If you want to
customize the plot and only display specific variables, you can provide their
names as a character vector. For example, setting `varx = c("var1", "var2")`
will plot only "var1" and "var2". Note that the variable names should match
the column names in the dataset.}

\item{truncate.lower, truncate.upper}{Specifies whether to truncate the
lower shrinkage factors to truncate.lower and/or the upper shrinkage factors
to truncate.upper. The default is NULL, which means no truncation is conducted.
It is important to note that truncation is performed before calculating the
averages and standard errors.}

\item{remove.zeros}{Specifies whether to remove zeros in the regression estimates
or shrinkage factors before calculating the averages and standard errors.
The default value is FALSE, indicating that zeros are not removed. If set to
TRUE, any zeros present in the regression estimates or shrinkage factors
will be excluded from the calculations. It is important to note that if
inclusion frequencies of variables are desired, zeros should not be removed.}

\item{linetype}{Specifies whether points with corresponding lines should be
plotted for the averages. The default is FALSE, which means only averages are
plotted. Please note that this option is only relevant when `plotype = "means"`.}

\item{add.se}{Specifies whether to include one standard error (SE) in
addition to the average values of the given metric. The default is FALSE. This
option is only relevant when `plotype = "means"`.}

\item{lwd, main, ylim, facet}{graphical parameters. see `[ggplot()]` and
`[plot]` for details}

\item{points_shape}{shape of points in the plot}

\item{points_size}{size of points in the plot}

\item{boxwidth}{boxplot width}
}
\description{
This function generates plots for the regression estimates, shrinkage factor,
or inclusion frequency of each variable. It provides the option to plot either
the means or distributions.
}
