% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_coefs_shrinkage}
\alias{plot_coefs_shrinkage}
\title{Plot regression estimate against their corresponding Shrinkage}
\usage{
plot_coefs_shrinkage(
  beta1,
  beta2,
  column = 1,
  remove.zeros = TRUE,
  points_shape = 16,
  points_size = 2,
  betatrue = 0,
  linewidth = 0.5,
  truncate.lower = -1,
  truncate.upper = 1.25
)
}
\arguments{
\item{beta1}{Matrix of regression estimates from a variable selection approach.}

\item{beta2}{Matrix of regression estimates from post-estimation shrinkage approach.}

\item{column}{Column that contains the variable of interest to be plotted.}

\item{remove.zeros}{Logical. Whether to remove rows with zero values in beta1 matrix.}

\item{points_shape}{Shape of the points in the plot (default: 16).}

\item{points_size}{Size of the points in the plot (default: 2).}

\item{betatrue}{True beta value for reference (default: 0).}

\item{linewidth}{Width of the dashed lines (default: 0.5).}

\item{truncate.lower}{A numeric value used to truncate negative shrinkage factors.
If a shrinkage factor is calculated to be less than this value, it will be set
to `truncate.lower`. Default is -1.}

\item{truncate.upper}{A numeric value used to truncate positive shrinkage factors.
If a shrinkage factor is calculated to be greater than this value, it will be set
to `truncate.upper`. Default is 1.25.}
}
\value{
A plot visualizing the relationship between shrinkage and regression estimates.
}
\description{
This function generates a plot to visualize the relationship between shrinkage
and regression estimates.
}
