% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_metric}
\alias{plot_metric}
\title{Plot the mean model error,relative test error, etc over several simulation settings.}
\usage{
plot_metric(
  x,
  betatypes = c("a", "b", "c", "d"),
  corrtypes = c("c1", "c2", "c3", "c4", "c5"),
  samplesize = NULL,
  snrx = NULL,
  nvar = 15,
  methods = NULL,
  xaxis = c("snr", "samplesize"),
  plotype = c("means", "median", "distribution"),
  legend.position = "right",
  row = c("betatype", "corrtype"),
  col = c("corrtype", "betatype"),
  linetype = FALSE,
  add.se = FALSE,
  lwd = 0.7,
  points_shape = 16,
  points_size = 2,
  dodgewidth = 0.5,
  main = NULL,
  ylim = NULL,
  ylab = NULL,
  xlab = NULL,
  boxwidth = 0.5,
  logx = FALSE,
  facet = FALSE,
  facet_size = 13,
  facet_color = "black",
  recode_methods = FALSE,
  new_methods_names = NULL,
  legend_title = NULL,
  add.xlab = TRUE,
  add.ylab = TRUE
)
}
\arguments{
\item{x}{The output of the `combine_sim_metrics()` function.}

\item{betatypes}{the beta type used in the simulation. Default is `betatypes = "a"`}

\item{corrtypes}{the correlation type type used in the simulation.
Default is `corrtypes = "a"`.}

\item{samplesize}{a vector of sample size used in the simulation. Default is NULL
and all the sample sizes used are plotted.}

\item{snrx}{a vector of signal to noise ratio used in the simulation. Default is NULL
and all the snr used are plotted.}

\item{nvar}{a numeric value for the number of variables used in the simulation.
Default is 15 variables}

\item{methods}{the name(s) of the method used in the simulation. Default is
null and all methods are plotted.}

\item{xaxis}{parameter to plot in the x axis. Default is `xaxis = snr`}

\item{plotype}{the plot to generate. Default is `plotype = "means"` which plots
the average values of a given metric. If distributions are required,
set `plotype = "distribution"` to generate boxplots.}

\item{legend.position}{the position of the legend in the plot. Default is to
place it on the right side of the plot}

\item{row, col}{One of `betatype` or `corrtype`,indicating the variables for
the rows and columns for plotting facet grid; note that row and col must be
different. Default is `row = betatype` and `col = corrtype`, so that the
plotting grid displays the given metric versus the SNR or sample size,
in a plotting grid with the coefficient types across the rows, and the
correlation levels across the columns.}

\item{linetype}{Specifies whether points with corresponding lines should be
plotted for the averages. The default is FALSE, which means only averages are
plotted. Please note that this option is only relevant when `plotype = "means"`.}

\item{add.se}{Specifies whether to include one standard error (SE) in
addition to the average values of the given metric. The default is FALSE. This
option is only relevant when `plotype = "means"`.}

\item{lwd, main, ylim, ylab, xlab, facet}{graphical parameters. see `[ggplot()]` and
`[plot]` for details}

\item{points_shape}{shape of points in the plot}

\item{points_size}{size of points in the plot}

\item{dodgewidth}{determines the amount of separation between the grouped elements in
boxplot}

\item{boxwidth}{boxplot width}

\item{logx}{specifies whether to log-transform the mean when generating the plot}

\item{facet_size}{Text size of the facet.}

\item{facet_color}{Text color of the facet.}

\item{recode_methods}{Specifies whether to recode the names of the methods based on the
user-supplied 'new_methods_names'}

\item{new_methods_names}{List of new names for the method which is used by recode() function
in dplyr. For example, if the old names are c("lasso(cv)", "lasso(aic)) we can change the names
to just cv and aic using this approach: list(`lasso(cv)` = "cv", `lasso(aic)` = "aic")}

\item{legend_title}{The name of the legend title. Suitable when new methods_name is used description}

\item{add.xlab, add.ylab}{specifies whether to add x and y labels}
}
\description{
Plot the mean model error,relative test error, etc over several simulation settings.
}
