% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_linpred.r
\name{posterior_linpred}
\alias{posterior_linpred}
\title{posterior distribution of the linear predictor}
\usage{
posterior_linpred(object, X, output = "both")
}
\arguments{
\item{object}{list object returned by brg() or brl()}

\item{X}{data matrix (new or current observations).}

\item{output}{character. Either "samples", "summary" or "both".}
}
\value{
An object of class \code{\link{list}}
}
\description{
This function allows to obtain samples and compute summary statistics from
the posterior distribution of the linear predictor.
}
\author{
Gwenael G.R. Leday
}
