% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predict.r
\name{posterior_predict}
\alias{posterior_predict}
\title{posterior predictive distribution}
\usage{
posterior_predict(object, X, output = "both")
}
\arguments{
\item{object}{list object returned by brg(), brl(), brgl()...}

\item{X}{data matrix (new or current observations).}

\item{output}{character. Either "samples", "summary" or "both".}
}
\value{
An object of class \code{\link{list}}
}
\description{
This function allows to obtain samples and compute summary statistics
from the posterior predictive distribution.
}
\author{
Gwenael G.R. Leday
}
