% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_selection.R
\name{predict.bestsubsetfit}
\alias{predict.bestsubsetfit}
\title{make predictions from a "bestsubsetfit" object.}
\usage{
\method{predict}{bestsubsetfit}(object, newx = NULL, ...)
}
\arguments{
\item{object}{Fitted \code{"bestsubsetfit"} model object}

\item{newx}{Matrix of new values for \code{x} at which predictions are to be
made. Must be a matrix.}

\item{...}{not used at the moment}
}
\description{
Similar to other predict methods, this functions predicts fitted values,
from a fitted \code{"bestsubsetfit"} object.
}
