% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.R
\name{predict.lassofit}
\alias{predict.lassofit}
\title{make predictions from a "lassofit" object.}
\usage{
\method{predict}{lassofit}(object, newx = NULL, ...)
}
\arguments{
\item{object}{Fitted \code{"lassofit"} model object}

\item{newx}{Matrix of new values for \code{x} at which predictions are to be
made. Must be a matrix.}

\item{...}{not used at the moment}
}
\description{
Similar to other predict methods, this functions predicts fitted values,
from a fitted \code{"lassofit"} object.
}
