% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classical_linear_models.R
\name{predict.linearmodel}
\alias{predict.linearmodel}
\title{Make predictions from a "linearmodel" object.}
\usage{
\method{predict}{linearmodel}(object, newx = NULL, ...)
}
\arguments{
\item{object}{A fitted \code{"linearmodel"} model object.}

\item{newx}{A matrix of new values for \code{x} at which predictions are to 
be made. It must be a matrix.}

\item{...}{not used at the moment}
}
\description{
Make predictions from linear models of class \code{"linearmodel"}. It is 
similar to other predict methods in R.
}
