% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonnegative_garrote.R
\name{predict.nngfit}
\alias{predict.nngfit}
\title{make predictions from "nngfit" object.}
\usage{
\method{predict}{nngfit}(object, newx = NULL, ...)
}
\arguments{
\item{object}{Fitted \code{"nngfit"} model object}

\item{newx}{Matrix of new values for \code{x} at which predictions are to be
made. Must be a matrix.}

\item{...}{not used at the moment}
}
\description{
Similar to other predict methods, this functions predicts fitted values,
from a fitted \code{"nngfit"} object.
}
