% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimation.R
\name{predict.pwsf}
\alias{predict.pwsf}
\title{make predictions from a "pwsf" object.}
\usage{
\method{predict}{pwsf}(object, newx = NULL, ...)
}
\arguments{
\item{object}{Fitted \code{"pwsf"} object}

\item{newx}{Matrix of new values for \code{x} at which predictions are to be
made. Must be a matrix.}

\item{...}{not used at the moment}
}
\description{
Similar to other predict methods, this functions predicts fitted values,
from a fitted \code{"pwsf"} object.
}
