% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{predict.ridgefit}
\alias{predict.ridgefit}
\title{make predictions from a "ridgefit" object.}
\usage{
\method{predict}{ridgefit}(object, newx = NULL, ...)
}
\arguments{
\item{object}{Fitted \code{"ridgefit"} model object}

\item{newx}{Matrix of new values for \code{x} at which predictions are to be
made. Must be a}

\item{...}{not used at the moment}
}
\description{
Similar to other predict methods, this functions predicts fitted values,
from a fitted \code{"ridgefit"} object.
}
