% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimation.R
\name{pwsf}
\alias{pwsf}
\title{Estimate parameter-wise shrinkage factors for variables}
\usage{
pwsf(
  x,
  y,
  standardize = FALSE,
  varx = NULL,
  nfolds = 10,
  foldid = NULL,
  nonnegative = FALSE,
  loocv = FALSE
)
}
\arguments{
\item{x}{A standardized matrix of predictors with dimensions nobs x nvars, where 
each row represents an observation vector. If not standardized, set standardize to
TRUE.}

\item{y}{A numeric, centered quantitative vector of responses. If not centered,
set `standardize` to TRUE; this will standardize x but center y.}

\item{standardize}{Specifies whether to standardize the predictors matrix 
\code{`x`} to have mean 0 and unit variance. If set to TRUE, \code{`x`} will be
 standardized. Additionally, the response variable \code{`y`} will be centered
 by subtracting its mean. This standardization step can be useful to ensure 
 that the predictors are on a comparable scale. By default, 
 \code{standardize=FALSE}, indicating that no standardization will be performed.
 This assumes that users have already standardized their data.}

\item{varx}{Names of the variables selected by a procedure. Used to subset 
the predictor matrix \code{`x`}. The default value is NULL, and all column 
names of x are used.}

\item{nfolds}{The number of cross-validation. if nfolds = n it returns
LOOCV. The default value is \code{nfolds = 10}, which conducts 10-fold CV.}

\item{foldid}{An optional vector of values between 1 and nfolds identifying
what fold each observation is in while conducting cross-validation.}

\item{nonnegative}{Specify whether nonnegative shrinkage factors should be 
estimated. Default is FALSE, meaning that negative shrinkage factors might be 
estimated.}

\item{loocv}{Specifies whether leave-one-out cross-validation (LOOCV) should be
used to estimate shrinkage factors. The default is set to FALSE, indicating that
k-fold cross-validation is performed.}
}
\value{
Returns a list with the following components:
\item{ShrunkenRegCoef}{Shrunken regression estimates for the selected variables,
with coefficients of zero assigned to the unselected variables.}
\item{ShrinkageFactors}{Shrinkage factors for the selected variables, with 
shrinkage factors of zero assigned to the unselected variables.}
\item{variables_selected}{The variables chosen for estimating shrinkage
factors (varx)}
\item{nvar}{The number of variables selected by the procedure (length of 
\code{"varx"}).}
\item{x}{The standardized design matrix of all predictor variables.}
\item{y}{The centered response variable.}
}
\description{
Estimates parameter-wise shrinkage factors and
corresponding shrunken regression estimates using k-fold cross-validation (CV)
or leave-one-out CV. Parameter-wise shrinkage (PWS) approach is an extension
of the global shrinkage proposed by Sauerbrei (1999). Unlike global
shrinkage, it shrinks regression coefficients differently and was originally
proposed for use after model selection. However, Kipruto and Sauerbrei
(2024) modified the approach by imposing non-negativity constraints on the 
shrinkage factors to allow its usage in both full and selected models.
}
