% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_selection.R
\name{regsubsetAICBIC}
\alias{regsubsetAICBIC}
\title{Select the best subset using an information criterion.}
\usage{
regsubsetAICBIC(
  x,
  y,
  method = c("exhaustive", "backward"),
  standardize = FALSE
)
}
\arguments{
\item{x}{The standardized matrix of predictors.}

\item{y}{A numeric centered quantitative vector of response.}

\item{method}{Specifies which method to use for variable selection: either
exhaustive search or backward elimination method. The default is exhaustive
search via leaps and bound algorithm.}

\item{standardize}{Specifies whether to standardize the predictors matrix
\code{`x`} to have mean 0 and unit variance. If set to TRUE, \code{`x`} will be
 standardized. Additionally, the response variable \code{`y`} will be centered
 by subtracting its mean. This standardization step can be useful to ensure
 that the predictors are on a comparable scale. By default,
 \code{standardize=FALSE}, indicating that no standardization will be performed.
 This assumes that users have already standardized their data.}
}
\value{
It returns a list with the following components:
\item{AIC:}{The AIC for each model size, where the first element corresponds
to the model with one variable, and so on.}
\item{BIC:}{The BIC for each model size, where the first element corresponds
to the model with one variable, and so on.}
\item{AIC.min:}{The minimum AIC}
\item{BIC.min:}{The mininum BIC}
\item{nvarAIC:}{The number of variables selected by smallest AIC}
\item{nvarBIC:}{The number of variables selected by smallest BIC}
}
\description{
Conduct variable selection via exhaustive search or backward elimination and
return the best model using an information criterion. The model with the smallest
AIC or BIC among models of different sizes is considered the best model.
}
