% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data_master.R
\name{residual_variance}
\alias{residual_variance}
\title{Estimate residual variance from OLS model}
\usage{
residual_variance(x, y)
}
\arguments{
\item{x}{A standardized matrix of predictors.}

\item{y}{A centered response variable.}
}
\value{
An estimate of the residual variance (`sigma2`) and regression 
estimates (`coefficients`) without intercept.
}
\description{
Calculates the residual variance (mean squared error) from a 
least-squares model using a standardized matrix of predictors and a centered
response variable.
}
