% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{ridge_aic_bic}
\alias{ridge_aic_bic}
\title{Helper function that fits ridge with AIC and BIC tuning parameters}
\usage{
ridge_aic_bic(x, y, lambda, nlambda, standardize, criterion, sigma2)
}
\arguments{
\item{x}{A standardized matrix of predictors of dimension nobs x nvars;
each row is an observation vector.If not standardized, set standardize to
TRUE.}

\item{y}{A numeric, centered quantitative vector of responses. If not centered,
set `standardize` to TRUE; this will standardize x but center y. See `standardize.response`
for more details.}

\item{lambda}{Optional user-supplied lambda sequence. Default is NULL, and
the program chooses its own sequence.}

\item{nlambda}{The number of lambda values. Default is 100}

\item{standardize}{Specifies whether to standardize the predictors matrix
\code{`x`} to have mean 0 and unit variance. If set to TRUE, \code{`x`} will be
 standardized. Additionally, the response variable \code{`y`} will be centered
 by subtracting its mean. This standardization step can be useful to ensure
 that the predictors are on a comparable scale. By default,
 \code{standardize = FALSE}, indicating that no standardization will be performed.
 This assumes that users have already standardized x and centered y so that the intercept
 is zero.}

\item{criterion}{The criterion used to select the tuning parameters. Available
options are: \code{criterion = "cv"} which perform cross-validation to select
the optimal tuning parameter, while \code{criterion = "aic"} and
\code{criterion = "bic"} select the tuning parameter based on the AIC and BIC,
respectively.}

\item{sigma2}{The residual variance obtained by fitting the full OLS model without
any regularization or feature selection. It is used for the calculation of AIC and BIC.
To compute these metrics, the residual variance needs to be provided via the
`sigma2` parameter.}
}
\description{
Helper function that fits ridge with AIC and BIC tuning parameters
}
