% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{ridge_cv}
\alias{ridge_cv}
\title{Helper function that fits ridge with cross-validation tuning parameters}
\usage{
ridge_cv(
  x,
  y,
  nfolds,
  foldid,
  lambda,
  nlambda,
  type.measure,
  parallel,
  standardize
)
}
\arguments{
\item{x}{A standardized matrix of predictors of dimension nobs x nvars;
each row is an observation vector.If not standardized, set standardize to
TRUE.}

\item{y}{A numeric, centered quantitative vector of responses. If not centered,
set `standardize` to TRUE; this will standardize x but center y. See `standardize.response`
for more details.}

\item{nfolds}{Number of folds for cross-validation. Default is 10. Smallest
value allowable is nfolds = 3.}

\item{foldid}{an optional vector of values between 1 and nfolds identifying
what fold each observation is in. If supplied, nfolds can be missing.}

\item{lambda}{Optional user-supplied lambda sequence. Default is NULL, and
the program chooses its own sequence.}

\item{nlambda}{The number of lambda values. Default is 100}

\item{type.measure}{Loss function to use for cross-validation. Currently,
two options are available. The default option is \code{type.measure = "mse"},
which corresponds to squared error. Another option is \code{type.measure = "mae"}
(mean absolute error).}

\item{parallel}{The program use \code{"[cv.glmnet()]"} function which supports
parallel computing. To make it work, users must register parallel beforehand.
see `glmnet` package for details}

\item{standardize}{Specifies whether to standardize the predictors matrix
\code{`x`} to have mean 0 and unit variance. If set to TRUE, \code{`x`} will be
 standardized. Additionally, the response variable \code{`y`} will be centered
 by subtracting its mean. This standardization step can be useful to ensure
 that the predictors are on a comparable scale. By default,
 \code{standardize = FALSE}, indicating that no standardization will be performed.
 This assumes that users have already standardized x and centered y so that the intercept
 is zero.}
}
\description{
Helper function that fits ridge with cross-validation tuning parameters
}
