% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{scaleFUN}
\alias{scaleFUN}
\title{Scale Function}
\usage{
scaleFUN(x)
}
\arguments{
\item{x}{A numeric value.}
}
\value{
A string representation of the input numeric value with two decimal places.
}
\description{
This function scales numeric values to a string representation with two decimal places.
}
