% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data_master.R
\name{sd_x}
\alias{sd_x}
\title{Standard Deviation of a Vector}
\usage{
sd_x(x)
}
\arguments{
\item{x}{A vector of numeric values. Missing data or `NA` will be removed.}
}
\description{
Calculates the standard deviation using 'n' instead of '(n-1)' in the denominator,
aligning with the conventions of the glmnet package.
}
