\name{internals}
\alias{w.test_safe}
\alias{t.test_safe}

\alias{[-methods}
\alias{names-methods}
\alias{length-methods}

\alias{[,ttest,ANY,missing-method}
  \alias{[,Prob0,ANY,missing-method}
  \alias{[,bFdr,ANY,missing-method}
  \alias{[,basicFdr,ANY,missing-method}
  \alias{[,eFdr,ANY,missing-method}
  \alias{[,estimate,ANY,missing-method}
  \alias{[,loss,ANY,missing-method}
  \alias{[,pseudoFdr,ANY,missing-method}
  \alias{[,sFdr,ANY,missing-method}
  \alias{[,sortableEstimate,ANY,missing-method}
  \alias{[,unsortableEstimate,ANY,missing-method}
  \alias{length,bFdr-method}
  \alias{length,basicFdr-method}
  \alias{length,eFdr-method}
  \alias{length,pseudoFdr-method}
  \alias{length,sFdr-method}
  \alias{names,Prob0-method}
  \alias{names,bFdr-method}
  \alias{names,basicFdr-method}
  \alias{names,eFdr-method}
  \alias{names,estimate-method}
  \alias{names,pseudoFdr-method}
  \alias{names,sFdr-method}

\title{Internal Functions and Methods}
\description{
These functions are for internal use and/or for upcoming packages or not yet documented.
}
\usage{
w.test_safe(x,...)
t.test_safe(x,...)
}
\keyword{internal}

