% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimation.R
\name{shrinkcoef}
\alias{shrinkcoef}
\title{Estimates parameter-wise shrinkage factors using k-fold
cross-validation.}
\usage{
shrinkcoef(
  x,
  y,
  varx = NULL,
  nfolds = 10,
  foldid = NULL,
  standardize = FALSE,
  nonnegative = FALSE
)
}
\arguments{
\item{x}{A standardized matrix of predictors with dimensions nobs x nvars, where 
each row represents an observation vector. If not standardized, set standardize to
TRUE.}

\item{y}{A numeric, centered quantitative vector of responses. If not centered,
set `standardize` to TRUE; this will standardize x but center y.}

\item{varx}{The names of variables selected by a selection procedure. The 
default value is NULL, and all column names of x are used.}

\item{nfolds}{The number of cross-validation. if nfolds = n it returns
LOOCV. Default is 10-fold CV}

\item{foldid}{An optional vector of values between 1 and nfolds identifying
what fold each observation is in while conducting cross-validation.}

\item{standardize}{Specifies whether to standardize the predictors matrix 
\code{`x`} to have mean 0 and unit variance. If set to TRUE, \code{`x`} will be
 standardized. Additionally, the response variable \code{`y`} will be centered
 by subtracting its mean. This standardization step can be useful to ensure 
 that the predictors are on a comparable scale. By default, 
 \code{standardize=FALSE}, indicating that no standardization will be performed.
 This assumes that users have already standardized their data.}

\item{nonnegative}{Specify whether nonnegative shrinkage factors should be 
estimated. Default is FALSE, meaning that negative shrinkage factors
might be estimated.}
}
\value{
Returns a list with the following components:
\item{ShrunkenRegCoef:}{Shrunken regression estimates for the  variables `varx` without an
intercept.}
\item{shrinkagefactors:}{Shrinkage factors for the variables `varx`.}
}
\description{
This function is a wrapper for pwfs(), which is used to estimate 
parameter-wise shrinkage factors based on the approach proposed by 
Sauerbrei (1999) and Kipruto and Sauerbei (2024).
}
