% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data_master.R
\name{sim_data}
\alias{sim_data}
\title{Generates simulation data.}
\usage{
sim_data(
  n = 100,
  nvar = 15,
  ntest = 1e+05,
  corrtype = c("c1", "c2", "c3", "c4", "c5"),
  betatype = c("a", "b", "c", "d"),
  withinSignalCorr = 0.8,
  withinNoiseCorr = 0.8,
  betweenSignalNoiseCorr = 0,
  snr = 1,
  standardize.response = FALSE
)
}
\arguments{
\item{n}{The number of training observations in the training data (sample size). Default is 100}

\item{nvar}{The number of predictors in training and testing dataset. Default is 15}

\item{ntest}{The number of observations in the testing data. Default is 100,000 observation}

\item{corrtype}{The type of correlation matrix used. Four options are available:\cr
\code{c1}: Correlation matrix from Houwelingen and Sauerbrei (2013)\cr
\code{c2}: AR1(0.3) autocorrelation\cr
\code{c3}: AR1(0.8) autocorrelation\cr
\code{c4}: Correlation from educational Body fat dataset
\code{c5}: Block correlation matrix\cr
Note that for AR1 autocorrelation, the correlation between the ith and jth
variables is defined as AR1(rho) = rho^|i-j|.}

\item{betatype}{Specifies the type of true regression coefficients. Please refer
to the simulation protocol for more detailed information. The nonzero coefficients
in the regression model are fixed at a constant value of 7.}

\item{withinSignalCorr, withinNoiseCorr, betweenSignalNoiseCorr}{parameters of block
correlation type (c5). See corr_type() for details. description}

\item{snr}{Specifies the desired signal-to-noise ratio (SNR), i.e., 
\eqn{SNR = \frac{var(mu)}{\sigma^2}} where \eqn{var(mu)} is the variance
of linear predictor and \eqn{\sigma^2} is the error variance. The error
variance is set so that the given SNR is achieved. Default is SNR = 1.}

\item{standardize.response}{Specifies whether the training response 
variable should be standardized. Default is FALSE and y is only centered by mean}
}
\value{
A list with the following components:
  \item{\code{x}}{Standardized training predictors}
  \item{\code{y}}{Centered (standardized) training response variable} 
  \item{\code{xtest}}{Standardized testing predictors}
  \item{\code{ytest}}{Centered testing response variable}
  \item{\code{Sigma}}{Covariance matrix used to generate the data}
  \item{\code{beta}}{True regression coefficients used to generate the data}
  \item{\code{sd_y}}{Standard deviation of the training outcome}
}
\description{
Generate a predictor matrix \code{x} and response vector \code{y} based 
on a specified setup. Two sets of predictor-response pairs are generated: one for 
training and one for testing. The training \code{x} is standardized to have a mean of 0 
and unit variance, while the training response is centered and standardized if necessary. 
The test data are then standardized using statistics derived from the training data.
}
\details{
The data model is: \eqn{Y \sim N(X\beta, \sigma^2 I)}. The predictor 
variables have covariance matrix `Sigma`
}
\references{
The simulation setup is based on the study titled "Comparison of variable selection 
procedures and investigation of the role of shrinkage: a simulation protocol" by 
[Kipruto and Sauerbrei(2022)].
}
