% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data_master.R
\name{sim_master}
\alias{sim_master}
\title{Master function for running simulations.}
\usage{
sim_master(
  n = 100,
  nvar = 15,
  ntest = 1e+05,
  reg.funs,
  nrep = 2000,
  seed = 472095,
  file = NULL,
  file.rep = 5,
  corrtype = c("c1", "c2", "c3", "c4", "c5"),
  betatype = c("a", "b", "c", "d"),
  withinSignalCorr = 0.8,
  withinNoiseCorr = 0.8,
  betweenSignalNoiseCorr = 0,
  snr = 1,
  nfolds = 10,
  foldid = NULL,
  standardize.response = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{n}{The number of observations in the training data(sample size). Default is 100 
observations}

\item{nvar}{The number of predictors in the training and testing data.}

\item{ntest}{The number of observations in the testing data. Default is 100,000 observations}

\item{reg.funs}{A list of functions representing the regression
procedures to be evaluated in the simulation. Each element of the
list must be a function that takes `x`, `y`, `foldid`, and `sigma2` 
(i.e., the training predictor matrix, training response vector, vector 
of values between 1 and the number of folds identifying what fold each 
observation is in while conducting cross-validation, and residual variance 
from the full model) as its only four (mandatory) arguments. It must return 
an object with associated \code{coef} and \code{predict} methods. The 
\code{coef} method must take `object` (the returned object) and return a 
vector of coefficients without an intercept. The \code{predict} method must 
take `object` and `newx` (the returned object and a new predictor matrix) 
and return a matrix of predictions.}

\item{nrep}{The number of repetitions of which to average the results. 
Default is 2000.}

\item{seed}{The overall random number generator set before repetitions begin 
(for the reproducibility of simulation results). The default is 472095.}

\item{file}{Name of the file to save the simulation results using saveRDS. If
set to NULL, no simulation results will be saved. The default value is NULL.}

\item{file.rep}{The number of repetitions after which intermediate results 
are saved. If file.rep is set to 0, it indicates that simulation results 
should be saved only at the end with no intermediate saving. The default is 5.}

\item{corrtype}{The type of correlation matrix used. Four options are available:
\code{c1}: Correlation matrix from Houwelingen and Sauerbrei (2013)\cr
\code{c2}: AR1(0.3) autocorrelation\cr
\code{c3}: AR1(0.8) autocorrelation\cr
\code{c4}: Correlation from educational Body fat dataset
\code{c5}: Block correlation structure\cr
Note that for AR1 autocorrelation, the correlation between the ith and jth
variables is defined as AR1(rho) = rho^|i-j|.}

\item{betatype}{Specifies the type of true regression coefficients. Please refer
to the simulation protocol for more detailed information. The nonzero coefficients
in the regression model are fixed at a constant value of 7.}

\item{withinSignalCorr, withinNoiseCorr, betweenSignalNoiseCorr}{parameters of block
correlation type (c5). See corr_type() for details.
variable should be standardized. Default is FALSE and y is only centered by mean}

\item{snr}{Specifies the desired signal-to-noise ratio (SNR), i.e., 
\eqn{SNR = \frac{var(mu)}{\sigma^2}} where \eqn{var(mu)} is the variance
of linear predictor and \eqn{\sigma^2} is the error variance. The error
variance is set so that the given SNR is achieved. Default is SNR = 1.}

\item{nfolds}{Number of folds for cross-validation. Default is 10. 
Smallest value allowable is nfolds = 3.}

\item{foldid}{An optional vector of values between 1 and `nfolds` identifying 
what fold each observation is in. If supplied, nfolds can be missing.}

\item{standardize.response}{Specifies whether the training response variable 
should be standardized. If set to TRUE, the model fitted will have both `x` 
and `y` standardized, but the coefficients will be back-transformed to the 
scale of standardized x.}

\item{verbose}{Specifies whether intermediate progress should be printed. Default is FALSE.}
}
\value{
A list with the following components: TODO
\item{err.train}{training error.}
\item{err.val}{validation error.}
\item{err.test}{test error.} 
\item{prop}{test proportion of variance explained}
\item{risk}{risk}\item{nzs}{number of selected nonzero coefficients.} \item{fpos}{number of false positives} \item{tpos}{number of true positives}
\item{fneg}{number of false negatives.}\item{tneg}{number of true negatives} \item{error.rate}{error rate}
\item{opt}{relative optimism (difference in test error and training error, divided by training error).}
\item{MCC}{Matthews correlation coefficient.}
\item{F1}{F1-score.}
Each of length N, where N is the number of regression methods under consideration
(the length of reg.funs). The ith element of each list is a matrix of
dimension nrep x m, where m = 1 is the number of tuning parameters inherent to
the ith method.
}
\description{
Run a set of simulations with the specified configuration.
}
\references{
Kipruto, E. and Sauerbrei, W. (2022). Comparison of variable 
selection procedures and investigation of the role of shrinkage in linear
regression-protocol of a simulation study in low-dimensional data. Plos one, 
17(10), p.e0271240.
}
\seealso{
\code{\link{sim_data}}
}
