% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{smedian}
\alias{smedian}
\title{Helper function that computes the sample median}
\usage{
smedian(x, conf.int = 0.95, na.rm = TRUE)
}
\arguments{
\item{x}{numeric vector from which NAs will be removed automatically}

\item{conf.int}{The coverage probability the outer quantiles should target.
When the default, 0.95, is used, the lower and upper quantiles computed are
0.025 and 0.975.}

\item{na.rm}{defaults to TRUE unlike built-in functions, so that by default
NAs are automatically removed}
}
\value{
returns the following summary statistics:
\item{Median}{Median value of x}
\item{Lower}{Lower quantile}
\item{Uppper}{Upper quantile}
}
\description{
computes the sample median and a selected pair of outer quantiles
having equal tail areas.  It is identical to the function
\code{smedian.hilow()} in the \code{Hmisc} package.
}
