% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{std_mean}
\alias{std_mean}
\title{Helper function used to calculate standard error of the mean}
\usage{
std_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{vector of numeric values. NA are removed before calculating the
statistic}

\item{na.rm}{a logical evaluating to TRUE or FALSE indicating whether NA values should be stripped
before computating standard deviation.}
}
\value{
the standard error. if length of x is one we assume standard
deviation is 0
}
\description{
Helper function used to calculate standard error of the mean
}
