% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.r
\name{summarize}
\alias{summarize}
\title{Compute summary statistics from samples}
\usage{
summarize(res)
}
\arguments{
\item{res}{list containing mcmc samples}
}
\value{
An object of class \code{\link{matrix}}
}
\description{
This function provides an eight-number summary of mcmc samples
}
\details{
This function returns the eight-number summary for each parameter samples.
}
\references{
Leday, G.G.R. et al. (2019)
}
\author{
Gwenael G.R. Leday
}
