% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_df.R
\name{trace_matrix}
\alias{trace_matrix}
\title{Trace of a Matrix}
\usage{
trace_matrix(A)
}
\arguments{
\item{A}{The input matrix.}
}
\value{
The sum of the diagonal elements of the matrix.
}
\description{
Calculates the trace of a given matrix.
}
\details{
This function checks if the input is a matrix and returns the sum of its diagonal
elements.
}
\examples{
A <- matrix(1:9, 3, 3)
trace_matrix(A)

}
