% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic.r
\name{waic}
\alias{waic}
\title{Watanabe-Akaike information criterion}
\usage{
waic(object, y, X)
}
\arguments{
\item{object}{list object returned by brg(), brl(), brgl()...}

\item{y}{response vector.}

\item{X}{data matrix.}
}
\value{
An object of class \code{\link{list}} with components:
\itemize{
 \item{}{\code{lppd} computed log-pointwise predictive density}
 \item{}{\code{p_waic} effective number of parameters (Gelman et al., 2014; Equ. (12))}
 \item{}{\code{elppd} estimate of expected log-pointwise predictive density}
 \item{}{\code{waic} WAIC criterion}
}
}
\description{
This function computes the Watanabe-Akaike information criterion also known as
the widely applicable information criterion.
}
\references{
Watanabe, S., Opper, M. (2010). Asymptotic equivalence of Bayes cross validation
and widely applicable information criterion in singular learning theory.
Journal of machine learning research, 11(12). 3571–3594.\cr
Gelman, A., Hwang, J., & Vehtari, A. (2014). Understanding predictive
information criteria for Bayesian models. Statistics and computing, 24(6), 997-1016.\cr
}
\author{
Gwenael G.R. Leday
}
