% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.r
\name{likelihood}
\alias{likelihood}
\title{Obtain (summary of) samples from the (log-)likelihood}
\usage{
likelihood(object, y, X, log = FALSE, output = "both")
}
\arguments{
\item{object}{list object returned by brg() or brl()}

\item{y}{response vector.}

\item{X}{data matrix.}

\item{log}{logical. If TRUE probabilities are given and summarized on the log scale.}

\item{output}{character. Either "samples", "summary" or "both".}
}
\value{
An object of class \code{\link{list}}
}
\description{
This function provides (summaries of) samples from the (log-)likelihood for each observation.
}
\references{
Leday, G.G.R. et al. (2019)
}
\author{
Gwenael G.R. Leday
}
