% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.r
\name{ridge}
\alias{ridge}
\title{Bayesian ridge regression}
\usage{
ridge(y, X, prior = "invGamma", a = 1e-05, b = 1e-05, mcmc = 5000L,
  burnin = 1000L, thin = 10L, verbose = TRUE, light = FALSE)
}
\arguments{
\item{y}{response vector of length n.}

\item{X}{n by p data matrix.}

\item{prior}{character. Either "bp", "ig", "ml" or "cpo". See Details.}

\item{a}{hyperparameter.}

\item{b}{hyperparameter.}

\item{mcmc}{integer. Number of desired samples.}

\item{burnin}{integer. Number of burn-in samples.}

\item{thin}{integer. Number of consecutive samples to skip.}

\item{verbose}{logical. Whether information on progress should be be printed.}

\item{light}{logical. If TRUE, only return a summary of the samples.}
}
\value{
An object of class \code{\link{list}}
}
\description{
This function implements a fast Gibbs algorithm for the Bayesian ridge regression model.
}
\details{
blablabla
}
\examples{
\dontrun{
# Generate data
set.seed(2019)
n <- 20
p <- 20
X <- matrix(rnorm(n*p), n, p)
trueBeta <- rep(0, p)
trueBeta[1:10] <- 1
y <- as.vector(X\%*\%trueBeta + rnorm(n, sd=0.5))

# Run gibbs sampler
res <- ridge(y, X)

# Extract summary
res$summary
}

}
\references{
Leday, G.G.R. et al. (2019)
}
\author{
Gwenael G.R. Leday
}
