\name{shrinkage-package}
\alias{shrinkage-package}
\alias{shrinkage}
\docType{package}

\title{
  Bayesian regression using shrinkage priors
}

\description{
  This packages implements several Bayesian regression models
}

\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}

\author{
Gwenael G.R. Leday <gwenael.leday@mrc-bsu.cam.ac.uk>

Maintainer: Gwenael G.R. Leday <gwenael.leday@mrc-bsu.cam.ac.uk>
}

\references{
  No references yet.
}

\keyword{Bayesian, regression, ridge}

\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
